+++ src/main/java/es/org/cxn/backapp/security/DefaultJwtUtils.java
@@ -1,5 +1,5 @@
 
-package es.org.cxn.backapp.service;
+package es.org.cxn.backapp.security;
 
 import java.time.Duration;
 import java.time.Instant;
