+++ src/main/java/es/org/cxn/backapp/service/impl/DefaultAddressService.java
@@ -22,13 +22,15 @@
  * SOFTWARE.
  */
 
-package es.org.cxn.backapp.service;
+package es.org.cxn.backapp.service.impl;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import es.org.cxn.backapp.exceptions.AddressServiceException;
 import es.org.cxn.backapp.model.persistence.PersistentCountryEntity;
 import es.org.cxn.backapp.repository.CountryEntityRepository;
+import es.org.cxn.backapp.service.AddressService;
+import es.org.cxn.backapp.service.UserService;
 
 import java.util.List;
