+++ src/main/java/es/org/cxn/backapp/service/impl/DefaultEmailService.java
@@ -1,4 +1,4 @@
-package es.org.cxn.backapp.service;
+package es.org.cxn.backapp.service.impl;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
@@ -8,6 +8,7 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
+import es.org.cxn.backapp.service.EmailService;
 import jakarta.mail.Message.RecipientType;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.InternetAddress;
