+++ src/main/java/es/org/cxn/backapp/service/impl/DefaultInvoiceService.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package es.org.cxn.backapp.service;
+package es.org.cxn.backapp.service.impl;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
@@ -31,6 +31,7 @@ import es.org.cxn.backapp.model.InvoiceEntity;
 import es.org.cxn.backapp.model.persistence.PersistentCompanyEntity;
 import es.org.cxn.backapp.model.persistence.PersistentInvoiceEntity;
 import es.org.cxn.backapp.repository.InvoiceEntityRepository;
+import es.org.cxn.backapp.service.InvoiceService;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
