+++ src/test/java/es/org/cxn/backapp/test/integration/controller/UserControllerIntegrationTest.java
@@ -41,7 +41,7 @@ import es.org.cxn.backapp.model.form.requests.UserChangePasswordRequest;
 import es.org.cxn.backapp.model.form.responses.AuthenticationResponse;
 import es.org.cxn.backapp.model.form.responses.UserDataResponse;
 import es.org.cxn.backapp.model.persistence.user.UserType;
-import es.org.cxn.backapp.service.DefaultUserService;
+import es.org.cxn.backapp.service.impl.DefaultUserService;
 import es.org.cxn.backapp.test.utils.UsersControllerFactory;
 import jakarta.mail.Session;
 import jakarta.mail.internet.MimeMessage;
