+++ src/test/java/es/org/cxn/backapp/test/unit/services/EmailServiceTest.java
@@ -16,7 +16,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.mail.javamail.JavaMailSender;
 
-import es.org.cxn.backapp.service.DefaultEmailService;
+import es.org.cxn.backapp.service.impl.DefaultEmailService;
 import jakarta.mail.Message.RecipientType;
 import jakarta.mail.MessagingException;
 import jakarta.mail.internet.InternetAddress;
