+++ src/test/java/es/org/cxn/backapp/test/unit/services/FederateStateServiceTest.java
@@ -26,8 +26,8 @@ import es.org.cxn.backapp.model.UserEntity;
 import es.org.cxn.backapp.model.persistence.PersistentFederateStateEntity;
 import es.org.cxn.backapp.model.persistence.user.PersistentUserEntity;
 import es.org.cxn.backapp.repository.FederateStateEntityRepository;
-import es.org.cxn.backapp.service.DefaultFederateStateService;
 import es.org.cxn.backapp.service.UserService;
+import es.org.cxn.backapp.service.impl.DefaultFederateStateService;
 
 /**
  * Unit tests for the {@link DefaultFederateStateService} class.
