+++ src/test/java/es/org/cxn/backapp/test/unit/services/InvoiceServiceTest.java
@@ -14,7 +14,7 @@ import es.org.cxn.backapp.exceptions.InvoiceServiceException;
 import es.org.cxn.backapp.model.persistence.PersistentCompanyEntity;
 import es.org.cxn.backapp.model.persistence.PersistentInvoiceEntity;
 import es.org.cxn.backapp.repository.InvoiceEntityRepository;
-import es.org.cxn.backapp.service.DefaultInvoiceService;
+import es.org.cxn.backapp.service.impl.DefaultInvoiceService;
 
 import java.time.LocalDate;
 import java.util.Collections;
