+++ src/test/java/es/org/cxn/backapp/test/unit/services/PaymentSheetServiceTest.java
@@ -39,7 +39,7 @@ import es.org.cxn.backapp.repository.PaymentSheetEntityRepository;
 import es.org.cxn.backapp.repository.RegularTransportRepository;
 import es.org.cxn.backapp.repository.SelfVehicleRepository;
 import es.org.cxn.backapp.repository.UserEntityRepository;
-import es.org.cxn.backapp.service.DefaultPaymentSheetService;
+import es.org.cxn.backapp.service.impl.DefaultPaymentSheetService;
 
 /**
  * Unit tests for the {@link DefaultPaymentSheetService} class.
